\name{BatterFourSeam}
\alias{BatterFourSeam}

\title{
  Shiny App for Illustrating Four-Seamer Rates Over Zone for a Batter
}
\description{
  Shiny app to illustrate rates associated with four-seam fastballs using 2015-2020 Statcast data for a given batter
}
\usage{
  BatterFourSeam()
}
\author{
  Jim Albert
}


