\name{BattingProfile}
\alias{BattingProfile}

\title{
  Shiny App for Graphing Spray Angle and Launch Angle
}
\description{
  Shiny app to display several graphs for understanding the spray angle and launch angle tendencies for a specific batter.
}
\usage{
  BattingProfile()
}
\author{
  Jim Albert
}


