\name{BetaBinomialABC}
\alias{BetaBinomialABC}

\title{
  Shiny App for Illustrating ABC for Beta-Binomial Model
}
\description{
  Shiny app to illustrate brushing for an Approximate Bayesian Computation inference about the precision parameter for a beta-binomial multilevel model
}
\usage{
  BetaBinomialABC()
}
\author{
  Jim Albert
}


