\name{BR_batting_league_avgs}
\alias{BR_batting_league_avgs}

\title{
  Baseball Reference Batting Averages for all Seasons
}
\description{
  Season to season batting averages for MLB from Baseball Reference
}
\usage{
  BR_batting_league_avgs
}
\format{
   A data frame with 152 observations on the following 28 variables.
  \describe{
  \item{Year}{season}
  \item{Tms}{number of teams}
  \item{X.Bat}{number at bat}
  \item{BatAge}{average batting age}
  \item{R.G}{runs per game}
  \item{G}{games}
  \item{PA}{plate appearances per game}
  \item{AB}{at-bats per game}
  \item{R}{runs per game}
  \item{H}{hits per game}
  \item{X2B}{doubles per game}
  \item{X3B}{triples per game}
  \item{HR}{home runs per game}
  \item{RBI}{runs-batted-in per game}
  \item{SB}{stolen bases per game}
  \item{CS}{caught stealing per game}
  \item{BB}{walks per game}
  \item{SO}{strikeouts per game}
  \item{BA}{batting average}
  \item{OBP}{on-base percentage}
  \item{SLG}{slugging percentage}
  \item{OPS}{ops measure}
  \item{TB}{total bases}
  \item{GDP}{grounded into double plays}
  \item{HBP}{hit by pitches}
  \item{SH}{sacrifice hits}
  \item{SF}{sacrifice flies}
  \item{IBB}{intentional walks}
  }
}
\source{
Baseball Reference
}


