\name{BR_pitching_league_avgs}
\alias{BR_pitching_league_avgs}

\title{
  Baseball Reference Pitching Averages for all Seasons
}
\description{
  Season to season pitching averages for MLB from Baseball Reference
}
\usage{
  BR_pitching_league_avgs
}
\format{
   A data frame with 151 observations on the following 33 variables.
  \describe{
  \item{Year}{season}
  \item{Tms}{number of teams}
  \item{X.P}{number of pitchers}
  \item{PAge}{average pitching age}
  \item{R.G}{runs per game}
  \item{ERA}{era}
  \item{PA}{plate appearances per game}
  \item{G}{number of games}
  \item{GF}{proportion of games finished}
  \item{CG}{proportion of complete games}
  \item{SHO}{proportion of complete game shutouts}
  \item{tSho}{proportion of team shutouts}
  \item{SV}{proportion of saves}
  \item{IP}{innings pitched per game}
  \item{H}{hits per game}
  \item{R}{runs per game}
  \item{ER}{earned runs per game}
  \item{HR}{home runs per game}
  \item{BB}{walks per game}
  \item{IBB}{intentional walks per game}
  \item{SO}{strikeouts per game}
  \item{HBP}{hit by pitches per game}
  \item{BK}{balks per game}
  \item{WP}{wild pitches per game}
  \item{BF}{batters faced per game}
  \item{WHIP}{average WHIP}
  \item{BAbip}{BA on balls in play}
  \item{H9}{hits per nine innings}
  \item{HR9}{home runs per nine innings}
  \item{BB9}{walks per nine innings}
  \item{SO9}{strikeouts per nine innings}
  \item{SO.W}{ratio of strikeouts to walks}
  \item{E}{average number of errors per game}
  }
}
\source{
Baseball Reference
}


