\name{BrushingBattingLocations}
\alias{BrushingBattingLocations}

\title{
  Shiny App for Displaying In-Play Hitting Locations
}
\description{
  Shiny app to visualize in-play field locations for specific batters for the 2021 season
}
\usage{
  BrushingBattingLocations()
}
\author{
  Jim Albert
}


