\name{BrushinginPlay3a}
\alias{BrushinginPlay3a}

\title{
  Shiny App for Brushing Launch Variable, Zone and Field Measurements for Pitch and Count Types
}
\description{
  Shiny app to illustrate brushing to explore relationships between  launch variables, zone locations and field locations for selected pitch and count types.
}
\usage{
  BrushinginPlay3a()
}
\author{
  Jim Albert
}


