\name{CareerTrajectoryBatting}
\alias{CareerTrajectoryBatting}

\title{
Shiny App for Plotting Two Career Trajectories for Two Batters of Interest
}
\description{
  Shiny app to graph career trajectories for two selected batters and measure of interest.
}
\usage{
  CareerTrajectoryBatting()
}
\author{
  Jim Albert
}


