\name{CareerTrajectoryFielding}
\alias{CareerTrajectoryFielding}

\title{
Shiny App for Plotting Two Career Trajectories for Two Fielders of Interest
}
\description{
  Shiny app to graph career trajectories for two selected fielders and measure of interest.
}
\usage{
  CareerTrajectoryFielding()
}
\author{
  Jim Albert
}


