\name{CareerTrajectoryPitching}
\alias{CareerTrajectoryPitching}

\title{
  Shiny App for Plotting Two Career Trajectories for Two Pitchers of Interest}

\description{
  Shiny app to graph career trajectories for two selected pitchers and measure of interest.
}
\usage{
  CareerTrajectoryPitching()
}
\author{
  Jim Albert
}


