\name{chadwick}
\alias{chadwick}

\title{
  Chadwick MLBAM ids
}
\description{
  Chadwick listing of MLBAM ids
}
\usage{
  chadwick
}
\format{
   A data frame with 89718 observations on the following 3 variables.
  \describe{
  \item{name_first}{first name of player}
  \item{name_last}{last name of player}
  \item{key_mlbam}{player's Statcast id}
  }
}
\source{
Chadwick Baseball Bureau Register
}


