\name{ClutchHitting}
\alias{ClutchHitting}

\title{
  Shiny App for Displaying Clutch Performance of a Hitter
}
\description{
  Shiny app that simulates the season runs values of a player assuming that events are independent of the bases/outs situation.
}
\usage{
  ClutchHitting()
}
\author{
  Jim Albert
}


