\name{ExpectedBA}
\alias{ExpectedBA}

\title{
  Shiny App for Graphing BA Over Launch Conditions
}
\description{
  Shiny app that shows counts of batted balls, counts of hits, and batting averages over subregions of launch angle and exit velocity.
}
\usage{
  ExpectedBA()
}
\author{
  Jim Albert
}


