\name{ExpectedHR}
\alias{ExpectedHR}

\title{
  Shiny App for Graphing Home Runs Over Launch Conditions
}
\description{
  Shiny app that shows counts of batted balls, counts of home runs, and home run rates over subregions of launch angle and exit velocity.
}
\usage{
  ExpectedBA()
}
\author{
  Jim Albert
}


