\name{ExpectedwOBA}
\alias{ExpectedwOBA}

\title{
  Shiny App for Graphing wOBA Over Launch Conditions
}
\description{
  Shiny app that shows counts of batted balls, counts of hits, and expected wOBA values over subregions of launch angle and exit velocity.
}
\usage{
  ExpectedwOBA()
}
\author{
  Jim Albert
}


