\name{FF_15_20}
\alias{FF_15_20}

\title{
  Four-Seam Fastball Statcast Data
}
\description{
  Statcast measurements on four-seam fastballs for 2015-2020 seasons
}
\usage{
  FF_15_20
}
\format{
   A data frame with 1,368,736 observations on the following 10 variables.
  \describe{
  \item{plate_x}{horizontal location of pitch}
  \item{plate_z}{vertical location of pitch}
  \item{description}{description of result of pitch}
  \item{p_throws}{side of the pitching arm}
  \item{stand}{side of the batter}
  \item{type}{indicates if the pitch is strike, ball or in-play}
  \item{events}{outcome of in-play result}
  \item{batter}{batter id}
  \item{pitcher}{pitcher id}
  \item{Season}{side of batter}
  }
}
\source{
Basball Savant
}


