\name{fg2020batting}
\alias{fg2020batting}

\title{
  2020 Fangraphs batting leaders
}
\description{
  Fangraphs batting leasders for the 2020 season
}
\usage{
  fg2020batting
}
\format{
   A data frame with 142 observations on the following 23 variables.
  \describe{
  \item{Name}{name of hitter}
  \item{Team}{name of team}
  \item{G}{games played}
  \item{PA}{plate appearances}
  \item{HR}{home runs}
  \item{R}{runs scored}
  \item{RBI}{runs batted in}
  \item{SB}{stolen bases}
  \item{BB_Pct}{walk rate}
  \item{K_Pct}{strikeout rate}
  \item{ISO}{isolated power}
  \item{BABIP}{batting average on balls in play}
  \item{AVG}{batting average}
  \item{OBP}{on-base percentage}
  \item{SLG}{slugging percentage}
  \item{wOBA}{weighted on-base percentage}
  \item{wRC}{weighted runs created plus}
  \item{EV}{mean exit velocity}
  \item{BsR}{base running runs above average}
  \item{Off}{offensive runs above average}
  \item{Def}{defensive runs above average}
  \item{WAR}{wins above replacement}
  \item{playerid}{player id}
  }
}
\source{
Fangraphs
}


