\name{game_info}
\alias{game_info}
\title{
  Game Ids for MLB Games
}
\description{
  Statcast game ids for games played in seasons 2008 through 2021
}
\usage{
  game_info
}
\format{
   A data frame with 36,960 observations on the following 2 variables.
  \describe{
  \item{game_info}{date of game}
  \item{game_pk}{Statcast game id}
}}
\source{
Baseball Savant
}


