\name{HomeRunLaunchVariables}
\alias{HomeRunLaunchVariables}

\title{
  Shiny App to Display Home Run Rates for a Region of Launch Variables
}
\description{
  Shiny app to illustrate in-play home run rates over values of launch variables and comparisons with predictions from previous seasons.
}
\usage{
  HomeRunLaunchVariables()
}
\author{
  Jim Albert
}


