\name{InPlayRatesSpray}
\alias{InPlayRatesSpray}

\title{
  Shiny App for In-Play Hit Rates Over Launch Conditions and Fielding
}
\description{
  Shiny app to visualize in-play hit rates over launch conditions over different fielding configurations for the 2021 season
}
\usage{
  InPlayRatesSpray()
}
\author{
  Jim Albert
}


