\name{InPlayRatesSpray2}
\alias{InPlayRatesSpray2}

\title{
  Shiny App for In-Play Hit Rates Over Launch Conditions for Two Seasons
}
\description{
  Shiny app to compare 2019 and 2021 seasons with respect to in-play hit rates over launch conditions.
}
\usage{
  InPlayRatesSpray2()
}
\author{
  Jim Albert
}


