\name{LocationValue}
\alias{LocationValue}

\title{
  Shiny App for Displaying Pitch Locations and Pitch Values for a 2021 Pitcher
}
\description{
  A Shiny app that displays pitch locations and pitch values over the zone for a selected pitch type for a single pitcher from the 2021 season.
}
\usage{
  LocationValue()
}
\author{
  Jim Albert
}


