\name{LogitHomeRunRates}
\alias{LogitHomeRunRates}

\title{
  Shiny App for Comparing Home Run Rates for Two Seasons
}
\description{
  Shiny app to compare two seasons with respect to in-play rates and home run rates over launch conditions.
}
\usage{
  LogitHomeRunRates()
}
\author{
  Jim Albert
}


