\name{OptimalPitchLocation}
\alias{OptimalPitchLocation}

\title{
  Shiny App to Find the Optimal Location of a Pitch
}
\description{
  Shiny app to illustrate the optimal location of a particular pitch type on a particular set of counts
}
\usage{
  OptimalPitchLocation()
}
\author{
  Jim Albert
}


