\name{PitchCount2}
\alias{PitchCount2}

\title{
  Shiny App for Illustrating Pitch Locations on Different Counts
}
\description{
  Shiny app to illustrate pitch locations for different pitch types and counts using 2019 Statcast data for pitchers
}
\usage{
  PitchCount2()
}
\author{
  Jim Albert
}


