\name{PitcherFourSeam}
\alias{PitcherFourSeam}

\title{
  Shiny App for Illustrating Four-Seamer Rates Over Zone for a Pitcher
}
\description{
  Shiny app to illustrate rates associated with four-seam fastballs using 2015-2020 Statcast data for a given pitcher
}
\usage{
  PitcherFourSeam()
}
\author{
  Jim Albert
}


