\name{pitchers_2019}
\alias{pitchers_2019}
\title{
  Statcast Ids for 2019 Pitchers
}
\description{
  Statcast ids for 20 pitchers who pitched in 2019 season
}
\usage{
  pitchers_2019
}
\format{
   A data frame with 20 observations on the following 2 variables.
  \describe{
  \item{Name}{name of pitcher}
  \item{SC_id}{Statcast id}
}}
\source{
Baseball Savant
}


