\name{PitchingTwoSeasons}
\alias{PitchingTwoSeasons}

\title{
  Shiny App for Exploring Pitching Measures for Two Seasons
}
\description{
  Shiny app to graph a pitching measure for pitchers who have thrown 100 innings in 2021 and 2022 seasons
}
\usage{
  PitchingTwoSeasons()
}
\author{
  Jim Albert
}


