\name{PredictHomeRuns}
\alias{PredictHomeRuns}

\title{
  Shiny App for Predicting Home Run Rates from a Previous Season
}
\description{
  Shiny app to illustrate predicting home run rates for future seasons from data from a previous season.
}
\usage{
  PredictHomeRuns()
}
\author{
  Jim Albert
}


