\name{PredictingBattingRates}
\alias{PredictingBattingRates}

\title{
  Shiny App for Illustrating Prediction of Batting Rates
}
\description{
  Shiny app to illustrate predicting hitting, strikeout or home run rates for the remainder of the season using a multilevel model.  Rates are per at-bat (AB).
}
\usage{
  PredictingBattingRates()
}
\author{
  Jim Albert
}


