\name{PredictingBattingRatesPA}
\alias{PredictingBattingRatesPA}

\title{
  Shiny App for Illustrating Prediction or Batting Rates
}
\description{
  Shiny app to illustrate predicting hitting, strikeout or home run rates for the remainder of the season using a multilevel model.  Rates are per plate appearance (PA).
}
\usage{
  PredictingBattingRates()
}
\author{
  Jim Albert
}


