\name{PredictiveMaxOfer}
\alias{PredictiveMaxOfer}

\title{
  Shiny App for Illustrating Predictive Checking
}
\description{
  Shiny app to illustrate predictive checking of a binomial model where the probability of success has a beta prior
}
\usage{
  PredictiveMaxOfer()
}
\author{
  Jim Albert
}


