\name{retro2019}
\alias{retro2019}

\title{
  2019 Retrosheet Data
}
\description{
  Retrosheet hitting data for individual plate appearances for 2019 season
}
\usage{
  retro2019
}
\format{
   A data frame with 191,973 observations on the following 5 variables.
  \describe{
  \item{GAME_ID}{game id}
  \item{BAT_ID}{batter id}
  \item{AB_FL}{indicator if an official at-bat}
  \item{EVENT_CD}{code for PA event}
  \item{INN_CT}{inning number}
}}
\source{
Retrosheet
}


