\name{sc2021_ip3}
\alias{sc2021_ip3}

\title{
  2021 Statcast In-Play Data
}
\description{
  Statcast measurements on balls in-play for 2021 season
}
\usage{
  sc2019_ip3
}
\format{
   A data frame with 121,108 observations on the following 7 variables.
  \describe{
  \item{Game_Date}{game date}
  \item{game_pk}{game code}
  \item{batter}{Statcast code of batter}
  \item{at_bat_number}{at-bat number during game}
  \item{estimated_ba}{estimated batting average from launch speed and angle}
  \item{estimated_woba}{estimated wOBA from launch speed and angle}
  \item{events}{outcome of ball put in-play}
  }
}
\source{
Basball Savant
}


