\name{sc2022_called}
\alias{sc2022_called}

\title{
  2022 Statcast Called Strike Data
}
\description{
  Statcast measurements on called strikes for 2022 season
}
\usage{
  sc2022_called
}
\format{
   A data frame with 366,959 observations on the following 7 variables.
  \describe{
  \item{plate_x}{horizontal location of pitch}
  \item{plate_z}{vertical location of pitch}
  \item{Call}{umpire call on pitch}
  \item{Count}{balls and strikes count}
  \item{Balls_Plus_Strikes}{sum of balls and strikes}
  \item{inning}{inning}
  \item{pitch_type}{code for pitch type}
  }
}
\source{
Baseball Savant
}


