\name{sc_pitchers_2019c}
\alias{sc_pitchers_2019c}

\title{
  2019 Statcast Pitch Data
}
\description{
  Statcast measurements on pitches for 2019 season
}
\usage{
  sc_pitchers_2019c
}
\format{
   A data frame with 732,473 observations on the following 8 variables.
  \describe{
  \item{SC_id}{mlb id of pitcher}
  \item{Name}{name of pitcher}
  \item{pitch_type}{pitch type}
  \item{balls}{number of balls in count}
   \item{strikes}{number of strikes in count}
   \item{stand}{batting side of batter}
   \item{plate_x}{pitch location - x}
   \item{plate_z}{pitch location - z}
}}
\source{
Baseball Savant
}


