\name{scip2019_new}
\alias{scip2019_new}

\title{
  2019 Statcast In-Play Data
}
\description{
  Statcast measurements on balls in-play for 2019 season
}
\usage{
  scip2019_new
}
\format{
   A data frame with 125,751 observations on the following 6 variables.
  \describe{
  \item{player_name}{name of hitter}
  \item{stand}{side of batter}
  \item{hc_x}{landing location of ball in play - x}
  \item{hc_y}{landing location of ball in play - y}
   \item{launch_angle}{launch angle (degrees)}
   \item{H}{number variable that is 1 for a hit, 0 otherwise}
}
\source{
Basball Savant
}


