\name{top30homerun}
\alias{top30homerun}

\title{
  Data on top 30 home run hitters
}
\description{
  Data on individual home runs hit by top 30 home run hitters in MLB history
}
\usage{
  top30homerun()
}
\format{
   A data frame with 17,304 observations on the following 22 variables.
  \describe{
  \item{Num}{cumulative HR count}
  \item{HR}{cumulative HR count}
  \item{HRyr}{cumulative HR count season}
  \item{HRgm}{cumulative HR count game}
  \item{Date}{date}
  \item{BatTeam}{batting team}
  \item{HomeAway}{home or away}
  \item{DefTeam}{defensive team}
  \item{Pitcher}{pitcher name}
  \item{Score}{game score}
  \item{Inn}{inning}
  \item{Out}{number of outs}
  \item{RoB}{runners on base}
  \item{RBI}{RBI on hit}
  \item{BOP}{batting position}
  \item{Pos}{fielding position}
  \item{WPA}{win probability added}
  \item{bWE}{win expectancy}
  \item{Play.Description}{description of play}
  \item{Player}{last name of player}
  \item{Birthdate}{birthdate}
  \item{Age}{age}
  }
}
\author{
  Jim Albert
}


