\name{twenty_seasons_counts}
\alias{twenty_seasons_counts}

\title{
  Twenty seasons of runs scored
}
\description{
  Tabulation of runs scored for the 2000-2019 seasons
}
\usage{
  twenty_seasons_counts
}
\format{
   A data frame with 2280 observations on the following 4 variables.
  \describe{
  \item{Season}{baseball season}
  \item{STATE}{runners on base and outs state}
  \item{O_RUNS.ROI}{runs in the remainder of the inning}
  \item{N}{frequency}
  }
}
\source{
Retrosheet
}


