\name{twentyyears_RA}
\alias{twentyyears_RA}

\title{
  Twenty seasons of runs advantage values
}
\description{
  Runs advantage calculations for the 2000-2019 seasons
}
\usage{
  twentyyears_RA
}
\format{
   A data frame with 480 observations on the following 5 variables.
  \describe{
  \item{Season}{baseball season}
  \item{Bases}{runners on base}
  \item{Outs}{number of outs}
  \item{Coef}{ordinal regression coefficient}
  \item{Bases_Score}{bases score}
  }
}
\source{
Retrosheet
}


