\name{twentyyears_RE}
\alias{twentyyears_RE}

\title{
  Twenty seasons of runs expectancies
}
\description{
  Runs expectancy calculations for the 2000-2019 seasons
}
\usage{
  twentyyears_RE
}
\format{
   A data frame with 480 observations on the following 6 variables.
  \describe{
  \item{Out}{number of outs}
  \item{Bases}{runners on base}
  \item{Mean}{mean runs scored in remainder of inning}
  \item{Season}{baseball season}
  \item{Prob}{probability of scoring in the remainder of the inning}
  \item{Prob2}{probability of scoring 2 or more runs in the remainder of the inning}
  }
}
\source{
Retrosheet
}


