\name{wOBA_Matchups}
\alias{wOBA_Matchups}

\title{
  Shiny App for Graphing Smoothed wOBA Matchup Values
}
\description{
  Shiny app to graph smoothed wOBA values of all batters that face a particular pitcher, or all pitchers that face a particular batter using Retrosheet play-by-play data from the 1960 through 2021 seasons.
}
\usage{
  wOBA_Matchups()
}
\author{
  Jim Albert
}


