\name{wOBA_Multilevel_Comparison}
\alias{wOBA_Multilevel_Comparison}

\title{
  Shiny App for Comparison of Multilevel wOBA Estimates
}
\description{
  Shiny app to graph multilevel estimates of wOBA across age for two Hall of Fame batters of estimates
}
\usage{
  wOBA_Multilevel_Comparison()
}
\author{
  Jim Albert
}


