\name{Predicting Batting Rates}
\alias{Predicting Batting Rates}

\title{
  Shiny App for Illustrating Predicting Batting Rates
}
\description{
  Shiny app to illustrate predicting hitting, strikeout or home run rates for the remainder of the season using a multilevel model
}
\usage{
  PredictingBattingRates()
}
\author{
  Jim Albert
}


